# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CIV.IQ Public API",
    "description": "Open REST API for normalized U.S. government data — federal and state legislators, bills, votes, campaign finance, lobbying, and ML-powered intelligence analysis. No API key required. Rate limited to 60 requests/minute per IP.\n\n## Rate limiting\n\nResponses carry the IETF RateLimit-Limit / RateLimit-Remaining / RateLimit-Reset / RateLimit-Policy headers (plus the legacy X-RateLimit-* trio); a 429 includes Retry-After.\n\n## Versioning & deprecation policy\n\nThe API is versioned in the URL path (/api/v1); every v1 response carries X-API-Version. Additive changes ship within v1 and are recorded at GET /v1/changelog. Breaking changes only ship as a new URL version. Before any version is retired, its responses will carry Deprecation and Sunset headers (RFC 8594) plus Link rel=\"successor-version\" at least 6 months before shutdown. No version is currently deprecated. Full policy: https://civdotiq.org/docs/api#versioning\n\n## Errors\n\nEvery error is structured JSON: v1 endpoints return V1Error (integer code + meta), non-v1 endpoints return ApiError (string code + metadata). Unknown API paths return a JSON 404, never HTML.",
    "version": "1.2.1",
    "license": {
      "name": "MIT",
      "url": "https://github.com/civdotiq/civ.iq/blob/main/LICENSE"
    },
    "contact": {
      "name": "CIV.IQ",
      "url": "https://civdotiq.org"
    }
  },
  "servers": [
    {
...
    }
  ],
  "paths": {
...
    "/intelligence/sector/{sector}/leaderboard": {
      "get": {
        "summary": "Sector donation leaderboard",
        "description": "Top recipients of campaign contributions from a specific industry sector.",
        "operationId": "getSectorLeaderboard",
        "tags": ["Intelligence"],
        "parameters": [
          {
            "name": "sector",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Industry sector name (e.g., Defense, Health, Energy)",
            "example": "Defense"
          },
          {
            "name": "chamber",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["house", "senate"]
            },
            "description": "Filter by chamber"
          },
          {
            "name": "party",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["D", "R", "I"]
            },
            "description": "Filter by party"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 50
            },
            "description": "Number of results"
          }
        ],
        "responses": {
          "200": {
            "description": "Sector leaderboard",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SectorLeaderboardResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiBadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ApiInternalError"
          }
        }
      }
    },
...
    "/search/policy-area": {
      "get": {
        "summary": "Policy area search",
        "description": "Cross-domain search by policy area: bills, regulations, spending, and committees.",
        "operationId": "searchPolicyArea",
        "tags": ["Search"],
        "parameters": [
          {
            "name": "policyArea",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Policy area name",
            "example": "Health"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Results per category"
          }
        ],
        "responses": {
          "200": {
            "description": "Policy area results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "policyArea": {
                      "type": "string"
                    },
                    "bills": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "regulations": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "spending": {
                      "type": "object",
                      "nullable": true
                    },
                    "committees": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    },
                    "metadata": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ApiBadRequest"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/ApiInternalError"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Meta": {
        "type": "object",
        "properties": {
          "apiVersion": {
            "type": "string",
            "example": "v1"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "source": {
            "type": "string"
          },
          "license": {
            "type": "string",
            "example": "MIT"
          },
          "documentation": {
            "type": "string",
            "format": "uri"
          }
        },
        "required": ["apiVersion", "timestamp"]
      },
...
      "V1Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "string"
              }
            },
            "required": ["code", "message"]
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
...
      "SectorLeaderboardResponse": {
        "type": "object",
        "properties": {
          "sector": {
            "type": "string"
          },
          "legislators": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "bioguideId": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "party": {
                  "type": "string"
                },
                "state": {
                  "type": "string"
                },
                "totalFromSector": {
                  "type": "number"
                },
                "percentOfTotal": {
                  "type": "number"
                }
              }
            }
          },
          "metadata": {
            "type": "object"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "description": "Error envelope returned by non-v1 /api endpoints (intelligence, search, geocode, mesh). v1 endpoints use V1Error.",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "NOT_FOUND",
                  "VALIDATION_ERROR",
                  "INTERNAL_ERROR",
                  "UNAUTHORIZED",
                  "FORBIDDEN",
                  "RATE_LIMITED",
                  "SERVICE_UNAVAILABLE",
                  "BAD_REQUEST",
                  "NO_DATA"
                ],
                "description": "Machine-readable error code"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "string",
                "description": "Optional technical detail"
              }
            },
            "required": ["code", "message"]
          },
          "metadata": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "requestId": {
                "type": "string"
              }
            }
          }
        },
        "required": ["success", "error"]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Operator bearer token (CACHE_WARM_SECRET). Required for internal observability endpoints under /v1."
      }
    },
    "responses": {
...
      "TooManyRequests": {
        "description": "Rate limit exceeded (60 requests/minute per IP)",
        "headers": {
          "RateLimit-Limit": {
            "schema": {
              "type": "integer"
            },
            "description": "Maximum requests per window (IETF RateLimit headers)"
          },
          "RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Remaining requests in current window"
          },
          "RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds until the window resets"
          },
          "RateLimit-Policy": {
            "schema": {
              "type": "string"
            },
            "description": "Quota policy, e.g. '60;w=60' (60 requests per 60s window)"
          },
          "X-RateLimit-Limit": {
            "schema": {
              "type": "integer"
            },
            "description": "Legacy: maximum requests per window"
          },
          "X-RateLimit-Remaining": {
            "schema": {
              "type": "integer"
            },
            "description": "Legacy: remaining requests in current window"
          },
          "X-RateLimit-Reset": {
            "schema": {
              "type": "integer"
            },
            "description": "Legacy: Unix millisecond timestamp when the window resets"
          },
          "Retry-After": {
            "schema": {
              "type": "integer"
            },
            "description": "Seconds to wait before retrying"
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/V1Error"
            }
          }
        }
      },
      "ApiBadRequest": {
        "description": "Invalid request parameters",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
      "ApiInternalError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiError"
            }
          }
        }
      },
...
    }
  },
...
}